requirements.txt
alembic==1.8.1
anyio==3.6.1
Authlib==0.15.5
autopep8==1.7.0
azure-common==1.1.28
azure-core==1.25.0
azure-identity==1.9.0
azure-mgmt-core==1.3.2
azure-mgmt-rdbms==10.1.0
azure-mgmt-resource==21.0.0
azure-mgmt-subscription==3.0.0
Babel==2.10.3
bcrypt==3.2.2
bidict==0.22.0
blinker==1.5
boto3==1.20.54
botocore==1.23.54
Brotli==1.0.9
certifi==2022.6.15
cffi==1.15.1
chardet==4.0.0
click==8.1.3
cryptography==3.4.8
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
eventlet==0.33.0
fastapi==0.79.0
Flask==2.1.3
Flask-Babel==2.0.0
Flask-Compress==1.12
Flask-Gravatar==0.5.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-Security-Too==4.1.5
Flask-SocketIO==5.2.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
h11==0.13.0
httpagentparser==1.9.3
httptools==0.4.0
idna==2.10
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
ldap3==2.9.1
Mako==1.2.1
MarkupSafe==2.1.1
msal==1.18.0
msal-extensions==0.3.1
msrest==0.7.1
oauthlib==3.2.0
paramiko==2.11.0
passlib==1.7.4
pgadmin4==6.12
Pillow==9.2.0
portalocker==2.5.1
psutil==5.9.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.9.2
PyJWT==2.4.0
PyNaCl==1.5.0
pyotp==2.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-engineio==4.3.4
python-jose==3.3.0
python-multipart==0.0.5
python-socketio==5.7.1
pytz==2021.3
PyYAML==6.0
qrcode==7.3.1
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.5.2
simplejson==3.17.6
six==1.16.0
sniffio==1.2.0
speaklater3==1.4
SQLAlchemy==1.4.40
sqlparse==0.4.2
sshtunnel==0.4.0
starlette==0.19.1
toml==0.10.2
typing_extensions==4.3.0
ua-parser==0.15.0
urllib3==1.26.11
user-agents==2.2.0
uvicorn==0.18.2
uvloop==0.16.0
watchfiles==0.16.1
websockets==10.3
Werkzeug==2.0.3
WTForms==3.0.1
